src/main/java/school/faang/user_service/validate/SkillValidate.java
package school.faang.user_service.validate;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;
import school.faang.user_service.dto.skill.SkillDto;
import school.faang.user_service.exception.DataValidationException;
import school.faang.user_service.repository.SkillRepository;

@Component
@RequiredArgsConstructor
public class SkillValidate {
    private final SkillRepository skillRepository;

    public void validatorSkills(SkillDto skillDto) {
        if (skillDto.getTitle().isBlank()) {
            throw new DataValidationException("The skill already exists");
        }

        if (skillRepository.existsByTitle(skillDto.getTitle())) {
            throw new DataValidationException("Skill with title" + skillDto.getTitle() + "already exists");
        }
    }

    public void validatorSkillsTitle(SkillDto skillDto) {
        if (skillDto.getTitle().isBlank()) {
            throw new DataValidationException("The skill already exists");
        }
    }
}
